+++ kompound/src/main/java/org/kaazing/nuklei/kompound/LocalEndpoint.java
@@ -16,21 +16,22 @@
 
 package org.kaazing.nuklei.kompound;
 
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.kaazing.nuklei.MessagingNukleus;
 import org.kaazing.nuklei.Nukleus;
 import org.kaazing.nuklei.concurrent.ringbuffer.mpsc.MpscRingBufferReader;
 import org.kaazing.nuklei.function.Mikro;
 import org.kaazing.nuklei.function.Proxy;
-import org.kaazing.nuklei.net.TcpManagerHeadersDecoder;
 import org.kaazing.nuklei.protocol.http.HttpDispatcher;
+import org.kaazing.nuklei.protocol.tcp.TcpManagerHeadersDecoder;
+
 import uk.co.real_logic.agrona.MutableDirectBuffer;
 import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
 import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
-
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Local endpoint for set of services
