+++ protocol/http/src/main/java/org/kaazing/nuklei/protocol/http/HttpDispatcher.java
@@ -16,18 +16,24 @@
 
 package org.kaazing.nuklei.protocol.http;
 
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Queue;
+import java.util.function.Supplier;
+import java.util.stream.IntStream;
+
 import org.kaazing.nuklei.function.Mikro;
-import org.kaazing.nuklei.net.TcpManagerHeadersDecoder;
-import org.kaazing.nuklei.net.TcpManagerTypeId;
 import org.kaazing.nuklei.protocol.Coordinates;
 import org.kaazing.nuklei.protocol.ProtocolUtil;
+import org.kaazing.nuklei.protocol.tcp.TcpManagerHeadersDecoder;
+import org.kaazing.nuklei.protocol.tcp.TcpManagerTypeId;
+
 import uk.co.real_logic.agrona.DirectBuffer;
 import uk.co.real_logic.agrona.MutableDirectBuffer;
 import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
-
-import java.util.*;
-import java.util.function.Supplier;
-import java.util.stream.IntStream;
 
 /**
  */
