+++ protocol/tcp/pom.xml
@@ -1,22 +1,4 @@
-<!--
-  Copyright 2014 Kaazing Corporation, All rights reserved.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
 
@@ -24,14 +6,13 @@
         <groupId>org.kaazing</groupId>
         <artifactId>nuklei-parent</artifactId>
         <version>develop-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>nuklei-http</artifactId>
+    <artifactId>nuklei-tcp</artifactId>
     <packaging>jar</packaging>
 
-    <name>Nuklei HTTP</name>
-    <description>HTTP protocol implementation designed to work with Kompound containers and Nuklei</description>
+    <name>Nuklei TCP</name>
+    <description>TCP protocol implementation designed to work with Kompound containers and Nuklei</description>
     <url>https://github.com/kaazing/nuklei.git</url>
 
     <scm>
@@ -50,9 +31,9 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>nuklei-tcp</artifactId>
-            <version>${project.version}</version>
+            <groupId>uk.co.real-logic</groupId>
+            <artifactId>Agrona</artifactId>
+            <version>0.1</version>
         </dependency>
         <dependency>
             <groupId>org.hamcrest</groupId>
@@ -73,9 +54,24 @@
             <scope>test</scope>
         </dependency>
         <dependency>
+            <!-- Needed for EL only, not AST -->
+            <groupId>org.kaazing</groupId>
+            <artifactId>k3po.lang</artifactId>
+            <version>[1.99.99.99,2.1.0.0)</version>
+        </dependency>
+        <dependency>
             <groupId>org.kaazing</groupId>
             <artifactId>k3po.junit</artifactId>
+<!--
             <version>2.0.0-alpha-3</version>
+-->
+            <version>develop-SNAPSHOT</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.kaazing</groupId>
+            <artifactId>specification.tcp</artifactId>
+            <version>develop-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -84,15 +80,11 @@
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-checkstyle-plugin</artifactId>
+                <artifactId>maven-enforcer-plugin</artifactId>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.12.4</version>
-                <configuration>
-                    <forkMode>always</forkMode>
-                </configuration>
+                <artifactId>maven-checkstyle-plugin</artifactId>
             </plugin>
             <plugin>
                 <groupId>org.kaazing</groupId>
@@ -113,18 +105,8 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-failsafe-plugin</artifactId>
-                <configuration>
-                    <skipTests>${skipITs}</skipTests>
-                </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>integration-test</goal>
-                            <goal>verify</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
         </plugins>
     </build>
+
 </project>
