+++ protocol/tcp/src/main/java/org/kaazing/nuklei/protocol/tcp/TcpAcceptor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kaazing.nuklei.net;
+package org.kaazing.nuklei.protocol.tcp;
 
 import org.kaazing.nuklei.NioSelectorNukleus;
 import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
