+++ protocol/tcp/src/main/java/org/kaazing/nuklei/protocol/tcp/TcpManager.java
@@ -14,21 +14,7 @@
  * limitations under the License.
  */
 
-package org.kaazing.nuklei.net;
-
-import org.kaazing.nuklei.MessagingNukleus;
-import org.kaazing.nuklei.NioSelectorNukleus;
-import org.kaazing.nuklei.Nuklei;
-import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
-import org.kaazing.nuklei.concurrent.ringbuffer.mpsc.MpscRingBufferWriter;
-import org.kaazing.nuklei.net.command.TcpCloseConnectionCmd;
-import org.kaazing.nuklei.net.command.TcpDetachCmd;
-import org.kaazing.nuklei.net.command.TcpLocalAttachCmd;
-import org.kaazing.nuklei.net.command.TcpRemoteAttachCmd;
-import uk.co.real_logic.agrona.BitUtil;
-import uk.co.real_logic.agrona.MutableDirectBuffer;
-import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
-import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
+package org.kaazing.nuklei.protocol.tcp;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
@@ -37,6 +23,21 @@ import java.nio.channels.Selector;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.IntSupplier;
+
+import org.kaazing.nuklei.MessagingNukleus;
+import org.kaazing.nuklei.NioSelectorNukleus;
+import org.kaazing.nuklei.Nuklei;
+import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
+import org.kaazing.nuklei.concurrent.ringbuffer.mpsc.MpscRingBufferWriter;
+import org.kaazing.nuklei.protocol.tcp.command.TcpCloseConnectionCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpDetachCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpLocalAttachCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpRemoteAttachCmd;
+
+import uk.co.real_logic.agrona.BitUtil;
+import uk.co.real_logic.agrona.MutableDirectBuffer;
+import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
+import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
 
 /**
  */
