+++ protocol/tcp/src/main/java/org/kaazing/nuklei/protocol/tcp/TcpManagerHeadersDecoder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kaazing.nuklei.net;
+package org.kaazing.nuklei.protocol.tcp;
 
 import java.nio.ByteOrder;
