+++ protocol/tcp/src/main/java/org/kaazing/nuklei/protocol/tcp/TcpManagerProxy.java
@@ -14,22 +14,23 @@
  * limitations under the License.
  */
 
-package org.kaazing.nuklei.net;
-
-import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
-import org.kaazing.nuklei.concurrent.ringbuffer.mpsc.MpscRingBufferWriter;
-import org.kaazing.nuklei.net.command.TcpCloseConnectionCmd;
-import org.kaazing.nuklei.net.command.TcpDetachCmd;
-import org.kaazing.nuklei.net.command.TcpLocalAttachCmd;
-import org.kaazing.nuklei.net.command.TcpRemoteAttachCmd;
-import uk.co.real_logic.agrona.BitUtil;
-import uk.co.real_logic.agrona.DirectBuffer;
-import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
-import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
+package org.kaazing.nuklei.protocol.tcp;
 
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
+
+import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
+import org.kaazing.nuklei.concurrent.ringbuffer.mpsc.MpscRingBufferWriter;
+import org.kaazing.nuklei.protocol.tcp.command.TcpCloseConnectionCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpDetachCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpLocalAttachCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpRemoteAttachCmd;
+
+import uk.co.real_logic.agrona.BitUtil;
+import uk.co.real_logic.agrona.DirectBuffer;
+import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
+import uk.co.real_logic.agrona.concurrent.UnsafeBuffer;
 
 /**
  * Interface for sending commands to a {@link TcpManager}
