+++ protocol/tcp/src/main/java/org/kaazing/nuklei/protocol/tcp/TcpSender.java
@@ -14,46 +14,54 @@
  * limitations under the License.
  */
 
-package org.kaazing.nuklei.net;
+package org.kaazing.nuklei.protocol.tcp;
+
+import java.nio.ByteBuffer;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.kaazing.nuklei.MessagingNukleus;
 import org.kaazing.nuklei.NioSelectorNukleus;
 import org.kaazing.nuklei.Nuklei;
 import org.kaazing.nuklei.concurrent.MpscArrayBuffer;
-import org.kaazing.nuklei.net.command.TcpCloseConnectionCmd;
+import org.kaazing.nuklei.protocol.tcp.command.TcpCloseConnectionCmd;
 
-import java.nio.channels.SelectionKey;
-import java.util.HashMap;
-import java.util.Map;
+import uk.co.real_logic.agrona.BitUtil;
+import uk.co.real_logic.agrona.MutableDirectBuffer;
+import uk.co.real_logic.agrona.concurrent.AtomicBuffer;
 
 /**
  */
-public class TcpReceiver
+public class TcpSender
 {
     private static final int MPSC_READ_LIMIT = 10;
 
     private final MessagingNukleus messagingNukleus;
-    private final NioSelectorNukleus selectorNukleus;
     private final Map<Long, TcpConnection> connectionsByIdMap;
+    private final ByteBuffer sendByteBuffer;
     private final MpscArrayBuffer<Object> tcpManagerCommandQueue;
-    private final MpscArrayBuffer<Object> tcpSenderCommandQueue;
+    private final MpscArrayBuffer<Object> tcpReceiverCommandQueue;
 
-    public TcpReceiver(
+    public TcpSender(
         final MpscArrayBuffer<Object> commandQueue,
+        final AtomicBuffer sendBuffer,
         final NioSelectorNukleus selectorNukleus,
         final MpscArrayBuffer<Object> tcpManagerCommandQueue,
-        final MpscArrayBuffer<Object> tcpSenderCommandQueue)
+        final MpscArrayBuffer<Object> tcpReceiverCommandQueue)
     {
         final MessagingNukleus.Builder builder = new MessagingNukleus.Builder()
             .nioSelector(selectorNukleus)
+            .mpscRingBuffer(sendBuffer, this::sendHandler, MPSC_READ_LIMIT)
             .mpscArrayBuffer(commandQueue, this::commandHandler, MPSC_READ_LIMIT);
 
-        this.selectorNukleus = selectorNukleus;
         this.tcpManagerCommandQueue = tcpManagerCommandQueue;
-        this.tcpSenderCommandQueue = tcpSenderCommandQueue;
+        this.tcpReceiverCommandQueue = tcpReceiverCommandQueue;
 
         messagingNukleus = builder.build();
         connectionsByIdMap = new HashMap<>();
+        byte[] sendByteArray = sendBuffer.byteArray();
+        sendByteBuffer = (sendByteArray != null) ? ByteBuffer.wrap(sendByteArray) : sendBuffer.byteBuffer().duplicate();
+        sendByteBuffer.clear();
     }
 
     public void launch(final Nuklei nuklei)
@@ -67,16 +75,12 @@ public class TcpReceiver
         {
             final TcpConnection connection = (TcpConnection)obj;
 
-            try
-            {
-                selectorNukleus.register(connection.channel(), SelectionKey.OP_READ, connection::onReadable);
-                connectionsByIdMap.put(connection.id(), connection);
+            connectionsByIdMap.put(connection.id(), connection);
 
-                connection.informOfNewConnection();
-            }
-            catch (final Exception ex)
+            // pass on to receiver so it can hook things up also
+            if (!tcpReceiverCommandQueue.write(connection))
             {
-                ex.printStackTrace(); // TODO: temp
+                throw new IllegalStateException("could not write to command queue");
             }
         }
         else if (obj instanceof TcpCloseConnectionCmd)
@@ -86,8 +90,34 @@ public class TcpReceiver
 
             if (null != connection)
             {
-                selectorNukleus.cancel(connection.channel(), SelectionKey.OP_READ);
-                connection.receiverClosed();
+                connection.senderClosed();
+                informTcpManagerOfClose(connection);
+            }
+        }
+    }
+
+    private void sendHandler(final int typeId, final MutableDirectBuffer buffer, final int offset, final int length)
+    {
+        if (TcpManagerTypeId.SEND_DATA == typeId)
+        {
+            final TcpConnection connection = connectionsByIdMap.get(buffer.getLong(offset));
+
+            final int messageOffset = offset + BitUtil.SIZE_OF_LONG;
+            sendByteBuffer.limit(messageOffset + length - BitUtil.SIZE_OF_LONG);
+            sendByteBuffer.position(messageOffset);
+
+            if (null != connection)
+            {
+                connection.send(sendByteBuffer);
+            }
+        }
+        else if (TcpManagerTypeId.CLOSE_CONNECTION == typeId)
+        {
+            final TcpConnection connection = connectionsByIdMap.remove(buffer.getLong(offset));
+
+            if (null != connection)
+            {
+                connection.senderClosed();
                 informTcpManagerOfClose(connection);
             }
         }
